requirements.txt
# Python
psycopg2-binary>=2.9.9
psycopg[c]>=2.9.9
Flask>=3.0.0
PyYAML>=6.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Dev
python-dotenv>=1.0.0
